developerContracting - Group project/requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
jwt==1.3.1
PyJWT==2.3.0 